requirements.txt
aiohttp
bs4
sqlalchemy
python-dotenv
psycopg2
SQLAlchemy-serializer
pydantic